requirements.txt
appnope==0.1.0
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
django-render-block==0.6
django-templated-email==2.3.0
django-widget-tweaks==1.4.5
gunicorn==19.9.0
idna==2.8
ipdb==0.12
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.3
wrapt==1.11.2
